moved vectorizing to trpc endpoints
