+++ src/components/PdfRenderer.tsx
@@ -1,35 +1,25 @@
 "use client";
 
 import { useState } from "react";
-import {
-  ChevronDown,
-  ChevronUp,
-  Loader2,
-  RotateCw,
-  Search,
-} from "lucide-react";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useToast } from "./ui/use-toast";
-import { useResizeDetector } from "react-resize-detector";
 import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod";
+import { ChevronDown, ChevronUp, Loader2 } from "lucide-react";
+import { useResizeDetector } from "react-resize-detector";
+import { Document, Page, pdfjs } from "react-pdf";
+import SimpleBar from "simplebar-react";
+import { cn } from "@/lib/utils";
+import { usePdfPagination } from "../lib/hooks/usePdfPagination";
+import { useToast } from "./ui/use-toast";
 import { Button } from "./ui/button";
 import { Input } from "./ui/input";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "./ui/dropdown-menu";
+import PdfFullScreen from "./PdfFullScreen";
+import PdfZoom from "./PdfZoom";
 
-import { Document, Page, pdfjs } from "react-pdf";
 import "react-pdf/dist/esm/Page/AnnotationLayer.css";
 import "react-pdf/dist/esm/Page/TextLayer.css";
-import { cn } from "@/lib/utils";
+import PdfRotate from "./PdfRotate";
 pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.js`;
-
-import SimpleBar from "simplebar-react";
-import PdfFullScreen from "./PdfFullScreen";
 
 interface PdfRendererProps {
   url: string;
@@ -39,20 +29,26 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
   const { toast } = useToast();
   const { width, ref } = useResizeDetector();
 
-  const [pagesCount, setPagesCount] = useState<number>();
-  const [currPage, setCurrPage] = useState(1);
   const [zoom, setZoom] = useState(1);
   const [rotation, setRotation] = useState(0);
   const [renderedScale, setRenderedScale] = useState<number | null>(null);
 
   const isLoading = renderedScale !== zoom;
 
-  const CustomPageValidator = z.object({
-    page: z
-      .string()
-      .refine((val) => Number(val) > 0 && Number(val) <= pagesCount!),
-  });
+  const {
+    handlePrevClick,
+    handleNextClick,
+    handlePageSubmit,
+    setPagesCount,
+    isPrevDisabled,
+    isNextDisabled,
+    currPage,
+    pagesCount,
+    CustomPageValidator,
+  } = usePdfPagination();
+
   type Page = z.infer<typeof CustomPageValidator>;
+
   const {
     register,
     handleSubmit,
@@ -63,11 +59,6 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
     resolver: zodResolver(CustomPageValidator),
   });
 
-  const handlePageSubmit = ({ page }: Page) => {
-    setCurrPage(Number(page));
-    setValue("page", page);
-  };
-
   return (
     <div className="w-full bg-white rounded-md shadow flex flex-col items-center">
       {/** TOOLBAR START */}
@@ -76,10 +67,10 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
         <div className="flex items-center gap-1.5">
           <Button
             onClick={() => {
-              setCurrPage((prev) => (prev - 1 > 1 ? prev - 1 : 1));
+              handlePrevClick();
               setValue("page", String(currPage - 1));
             }}
-            disabled={currPage <= 1}
+            disabled={isPrevDisabled}
             variant="ghost"
             aria-label="previous page"
           >
@@ -90,9 +81,9 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
             <Input
               {...register("page")}
               onKeyDown={(e) => {
-                console.log(e.key);
                 if (e.key === "Enter") {
                   handleSubmit(handlePageSubmit)();
+                  setValue("page", String(currPage));
                 }
               }}
               className={cn(
@@ -108,12 +99,10 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
 
           <Button
             onClick={() => {
-              setCurrPage((prev) =>
-                prev + 1 > pagesCount! ? pagesCount! : prev + 1
-              );
+              handleNextClick();
               setValue("page", String(currPage + 1));
             }}
-            disabled={!pagesCount || currPage === pagesCount}
+            disabled={isNextDisabled}
             variant="ghost"
             aria-label="next page"
           >
@@ -122,38 +111,8 @@ const PdfRenderer = ({ url }: PdfRendererProps) => {
         </div>
         {/** TOOLS */}
         <div className="space-x-2">
-          {/** ZOOM */}
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button aria-label="zoom" variant="ghost" className="gap-1.5">
-                <Search className="h-4 w-4" />
-                {zoom * 100}%<ChevronDown className="h-3 w-3 opacity-50" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent>
-              <DropdownMenuItem onSelect={() => setZoom(1)}>
-                100%
-              </DropdownMenuItem>
-              <DropdownMenuItem onSelect={() => setZoom(1.5)}>
-                150%
-              </DropdownMenuItem>
-              <DropdownMenuItem onSelect={() => setZoom(2)}>
-                200%
-              </DropdownMenuItem>
-              <DropdownMenuItem onSelect={() => setZoom(2.5)}>
-                250%
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          {/** ROTATE */}
-          <Button
-            aria-label="rotate 90 degrees"
-            variant="ghost"
-            onClick={() => setRotation((prev) => prev + 90)}
-          >
-            <RotateCw className="h-4 w-4" />
-          </Button>
-          {/** FULL SCREEN */}
+          <PdfZoom zoom={zoom} setZoom={setZoom} />
+          <PdfRotate setRotation={setRotation} />
           <PdfFullScreen fileUrl={url} />
         </div>
       </div>
