+++ src/components/UploadButton.tsx
@@ -1,194 +1,9 @@
 "use client";
 
 import { useState } from "react";
-import { Dialog, DialogContent } from "./ui/dialog";
-import { DialogTrigger } from "@radix-ui/react-dialog";
+import { Dialog, DialogTrigger, DialogContent } from "./ui/dialog";
 import { Button } from "./ui/button";
-import Dropzone from "react-dropzone";
-import { Cloud, File, Loader2 } from "lucide-react";
-import { Progress } from "./ui/progress";
-import { useToast } from "./ui/use-toast";
-import { trpc } from "@/app/_trpc/client";
-import { useRouter } from "next/navigation";
-import { PDFLoader } from "langchain/document_loaders/fs/pdf";
-import { OpenAIEmbeddings } from "langchain/embeddings/openai";
-import { PineconeStore } from "langchain/vectorstores/pinecone";
-import { getPinecone } from "@/lib/pinecone";
-import { db } from "@/db";
-
-const UploadDropzone = () => {
-  const router = useRouter();
-  const { toast } = useToast();
-
-  const [isUploading, setIsUploading] = useState(false);
-  const [uploadProgress, setUploadProgress] = useState(0);
-  const [file, setFile] = useState<File | null>(null);
-
-  const { mutate: startPolling } = trpc.getFile.useMutation({
-    onSuccess: (file) => {
-      router.push(`/dashboard/${file.id}`);
-    },
-    retry: true,
-    retryDelay: 500,
-  });
-
-  const startSimulatedProgress = () => {
-    setUploadProgress(0);
-
-    const interval = setInterval(() => {
-      setUploadProgress((prev) => {
-        if (prev >= 95) {
-          clearInterval(interval);
-          return prev;
-        }
-        return prev + 5; // increment progress in 5% step
-      });
-    }, 500);
-
-    return interval;
-  };
-
-  const { mutate: saveFile } = trpc.createFile.useMutation({
-    onSuccess: (file) => {
-      console.log(file);
-    },
-  });
-  const { data: signedUrl } = trpc.createSignedUrl.useQuery();
-
-  return (
-    <Dropzone
-      multiple={false}
-      onDrop={async (acceptedFiles) => {
-        setIsUploading(true);
-        const progressInterval = startSimulatedProgress();
-        setFile(acceptedFiles[0]);
-
-        if (!signedUrl?.key) {
-          return toast({
-            title: "Something went wrong",
-            description: "Please try again later",
-            variant: "destructive",
-          });
-        }
-
-        const formData = new FormData();
-        formData.append("file", acceptedFiles[0]);
-
-        await fetch(signedUrl.url, {
-          method: "PUT",
-          body: formData,
-        });
-
-        const createdFile = saveFile({
-          key: signedUrl.key,
-          name: acceptedFiles[0].name,
-          url: `https://stream-bucket1.s3.eu-north-1.amazonaws.com/${signedUrl.key}`,
-        });
-        console.log("created", createdFile);
-        try {
-          const res = await fetch(
-            `https://stream-bucket1.s3.eu-north-1.amazonaws.com/${signedUrl.key}`
-          );
-          const blob = await res.blob();
-
-          const loader = new PDFLoader(blob);
-          const pageLevelDocs = await loader.load();
-
-          const pagesCount = pageLevelDocs.length;
-
-          //vectorize and index the entire document
-          const pineconeIndex = getPinecone();
-          const embeddings = new OpenAIEmbeddings({
-            openAIApiKey: process.env.OPENAI_API_KEY,
-          });
-
-          await PineconeStore.fromDocuments(pageLevelDocs, embeddings, {
-            pineconeIndex,
-          });
-
-          await db.file.update({
-            data: {
-              uploadStatus: "SUCCESS",
-            },
-            where: {
-              id: createdFile.id,
-            },
-          });
-        } catch (err) {
-          await db.file.update({
-            data: {
-              uploadStatus: "FAILED",
-            },
-            where: {
-              id: createdFile.id,
-            },
-          });
-        }
-        clearInterval(progressInterval);
-        setUploadProgress(100);
-        startPolling({ key: signedUrl.key });
-      }}
-    >
-      {({ getRootProps, getInputProps, acceptedFiles }) => (
-        <div
-          {...getRootProps()}
-          className="border h-64 m-4 border-dashed border-gray-300 rounded-lg"
-        >
-          <div className="flex items-center justify-center h-full w-full">
-            <label
-              htmlFor="dropzone-file"
-              className="flex flex-col items-center justify-center w-full h-full rounded-lg cursor-pointer bg-gray-50 hover:bg-gray-100"
-            >
-              <div className="flex flex-col items-center justify-center pt-5 pb-6">
-                <Cloud className="h-6 w-6 text-zinx-500 mb-2" />
-                <p className="mb-2 text-sm text-zinc-700">
-                  <span className="font-semibold">Click to upload</span> or grad
-                  and drop
-                </p>
-                <p className="text-xs text-zinc-500">PDF (up to 4MB)</p>
-              </div>
-              {acceptedFiles && acceptedFiles[0] ? (
-                <div className="max-w-xs bg-white flex items-center rounded-md overflow-hidden outline outline-[1px] outline-zinc-200 divide-x divide-zinc-200">
-                  <div className="px-3 py-2 h-full grid place-items-center">
-                    <File className="h-4 w-4 text-blue-500" />
-                  </div>
-                  <div className="px-3 py-2 h-full text-sm truncate">
-                    {acceptedFiles[0].name}
-                  </div>
-                </div>
-              ) : null}
-
-              {isUploading ? (
-                <div className="w-full max-w-xs max-auto mt-4">
-                  <Progress
-                    indicatorColor={
-                      uploadProgress === 100 ? "bg-green-500" : ""
-                    }
-                    value={uploadProgress}
-                    className="h-1 w-fill bg-zinc-200"
-                  />
-                  {uploadProgress === 100 ? (
-                    <div className="flex gap-1 items-center justify-center text-sm text-zinc-700 text-center pt-2">
-                      <Loader2 className="h-2 w-3 animate-spin" />
-                      Redirecting...
-                    </div>
-                  ) : null}
-                </div>
-              ) : null}
-
-              <input
-                type="file"
-                id="dropzone-file"
-                className="hidden"
-                {...getInputProps()}
-              />
-            </label>
-          </div>
-        </div>
-      )}
-    </Dropzone>
-  );
-};
+import UploadDropzone from "./UploadDropZone";
 
 const UploadButton = () => {
   const [isOpen, setIsOpen] = useState(false);
